Added DashboardUtils.js and removed unused dependency
