+++ test-runner-frontend/package.json
@@ -17,7 +17,6 @@
     "buffer": "^6.0.3",
     "firebase": "^9.9.1",
     "firebase-admin": "^11.0.1",
-    "jspdf-html2canvas": "^1.4.9",
     "react": "^18.2.0",
     "react-bootstrap": "^2.4.0",
     "react-dom": "^18.2.0",
